requirements.txt
Django==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
channels==4.0.0
python-decouple==3.8
